+++ tests/unit/custom-granularity-calendar.spec.ts
@@ -57,7 +57,7 @@ describe('CustomGranularityCalendar', () => {
             `${currentYear}-02-01T00:00:00.000Z`,
           );
           expect(emittedDate.end?.toISOString()).toStrictEqual(
-            `${currentYear}-03-01T00:00:00.000Z`,
+            `${currentYear}-02-28T23:59:59.999Z`,
           );
           expect(emittedDate.duration).toBe('month');
         });
@@ -137,19 +137,17 @@ describe('CustomGranularityCalendar', () => {
       const Q2 = options[1];
       expect(RANGE_PICKERS.quarter.selectableRanges.optionToRange(Q2)).toStrictEqual({
         start: new Date(Date.UTC(2016, 3, 1)),
-        end: new Date(Date.UTC(2016, 6, 1)),
+        end: new Date(Date.UTC(2016, 5, 30, 23, 59, 59, 999)),
         duration: 'quarter',
       });
     });
 
     it('should convert a arbitrary date range to the associated option', () => {
       expect(
-        RANGE_PICKERS.quarter.selectableRanges.rangeToOption({
+        RANGE_PICKERS.quarter.selectableRanges.rangeToOption(
           // Weird quarter from the 12th to the 12th
-          start: new Date(Date.UTC(2016, 3, 12)),
-          end: new Date(Date.UTC(2016, 6, 12)),
-          duration: 'quarter',
-        }),
+          new Date(Date.UTC(2016, 3, 12)),
+        ),
       ).toStrictEqual(DateTime.utc(2016, 4, 1, { locale: 'en' }));
     });
   });
@@ -182,19 +180,17 @@ describe('CustomGranularityCalendar', () => {
       const Year2012 = options[2];
       expect(RANGE_PICKERS.year.selectableRanges.optionToRange(Year2012)).toStrictEqual({
         start: new Date(Date.UTC(2012, 0, 1)),
-        end: new Date(Date.UTC(2013, 0, 1)),
+        end: new Date(Date.UTC(2012, 11, 31, 23, 59, 59, 999)),
         duration: 'year',
       });
     });
 
     it('should convert a arbitrary date range to the associated option', () => {
       expect(
-        RANGE_PICKERS.year.selectableRanges.rangeToOption({
+        RANGE_PICKERS.year.selectableRanges.rangeToOption(
           // Weird year from the 12th feb to the 12th feb
-          start: new Date(Date.UTC(2012, 1, 12)),
-          end: new Date(Date.UTC(2013, 1, 12)),
-          duration: 'year',
-        }),
+          new Date(Date.UTC(2012, 1, 12)),
+        ),
       ).toStrictEqual(DateTime.utc(2012, 1, 1, { locale: 'en' }));
     });
   });